# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: 'LBAW FinJet Web API'
  description: 'Recursos web (A7) para FinJet'
servers:
- url: http://lbaw2191.lbaw.fe.up.pt
  description: 'Servidor de Produção'
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Formulário de Login'
      description: 'Providencia um formulário de login. Acesso: PUB'
      tags:
      - 'M01: Autenticação e perfil individual'
      responses:
        '200':
          description: 'Ok. Mostra formulário de login'
    post:
      operationId: R102
      summary: 'R102: Ação de Login'
      description: 'Processa a submissão do formulário de login. Acesso: PUB'
      tags:
        - 'M01: Autenticação e perfil individual'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email: 
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '302': 
          description: 'Redirecionar depois de processar as credenciais de login'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Sucesso:
                  description: 'Autenticação sucedida. Redirecção para a homepage'
                  value: '/'
                302SucessoBanido:
                  description: 'Autenticação sucedida. Utilizador banido.'
                  value: '/perfil/{nomeUtilizador}'
                302Erro:
                  description: 'Autenticação falhou. Redirecção para a página de login'
                  value: '/login'
...
  /perfil/{nomeUtilizador}:
    get:
      operationId: R108
      summary: 'R108: Perfil do utilizador'
      description: 'Mostra a informação do perfil de um utilizador. Acesso: USR, OWN'
      tags:
        - 'M01: Autenticação e perfil individual'
      
      parameters:
        - in: path
          name: nomeUtilizador
          schema:
            type: string
          required: true
      responses:
        '200':
          description: 'Ok. Mostra o perfil do utilizador'
        '302':
          description: 'O utilizador autenticado não tem permissões suficientes. Por isso é redirecionado à página de login.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Sucesso:
                  description: 'Redireção à página de login'
                  value: '/login'
        '404':
          description: 'ID não existe. Mostra página de erro'
...